refactor: 유저 데이터 쿼리의 staleTime을 Infinity로 설정하여 캐싱 전략 개선
 - 유저 데이터는 로그인한 동안 변경되지 않는 경우가 많으므로 staleTime: Infinity 설정
 - 유저 데이터 수정, 삭제, 로그아웃에서 별도로 쿼리 무효화가 설정 된 상태
